requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org
--trusted-host nexus.ny1.ninetyone.com

aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.2
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
brotli==1.1.0
    # via yt-dlp
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   python-certifi-win32
    #   requests
    #   yt-dlp
charset-normalizer==3.3.2
    # via requests
clarabel==0.9.0
    # via cvxpy
colorama==0.4.6
    # via tqdm
coloredlogs==15.0.1
    # via onnxruntime
cvxpy==1.6.0
    # via pyportfolioopt
datasets==3.1.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
feedparser==6.0.11
    # via -r requirements.in
filelock==3.15.4
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
flatbuffers==24.3.25
    # via onnxruntime
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.6.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
furl==2.1.3
    # via -r requirements.in
gliner==0.2.13
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
huggingface-hub==0.24.6
    # via
    #   datasets
    #   gliner
    #   sentence-transformers
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via torch
jiter==0.5.0
    # via openai
joblib==1.4.2
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mutagen==1.47.0
    # via yt-dlp
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   clarabel
    #   cvxpy
    #   datasets
    #   onnxruntime
    #   osqp
    #   pandas
    #   pyportfolioopt
    #   qdldl
    #   scikit-learn
    #   scipy
    #   scs
    #   sentence-transformers
    #   transformers
onnxruntime==1.20.0
    # via gliner
openai==1.42.0
    # via
    #   -r requirements.in
    #   openai-messages-token-helper
openai-messages-token-helper==0.1.10
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
orderedmultidict==1.0.1
    # via furl
osqp==0.6.7.post3
    # via cvxpy
packaging==24.1
    # via
    #   datasets
    #   huggingface-hub
    #   onnxruntime
    #   setuptools-scm
    #   transformers
pandas==2.2.3
    # via
    #   datasets
    #   pyportfolioopt
pillow==10.4.0
    # via
    #   -r requirements.in
    #   openai-messages-token-helper
    #   sentence-transformers
pip-system-certs==4.0
    # via -r requirements.in
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
protobuf==5.28.3
    # via onnxruntime
pyarrow==18.0.0
    # via datasets
pycryptodomex==3.20.0
    # via yt-dlp
pydantic==2.8.2
    # via openai
pydantic-core==2.20.1
    # via
    #   -r requirements.in
    #   pydantic
pydub==0.25.1
    # via -r requirements.in
pygments==2.18.0
    # via rich
pyodbc==5.2.0
    # via -r requirements.in
pyportfolioopt==1.5.5
    # via -r requirements.in
pyreadline3==3.5.4
    # via humanfriendly
python-certifi-win32==1.6.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
qdldl==0.1.7.post4
    # via osqp
redis==5.0.8
    # via spotipy
regex==2024.7.24
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   spotipy
    #   tiktoken
    #   transformers
    #   yt-dlp
rich==13.9.4
    # via -r requirements.in
safetensors==0.4.4
    # via transformers
scikit-learn==1.5.1
    # via sentence-transformers
scipy==1.14.1
    # via
    #   -r requirements.in
    #   clarabel
    #   cvxpy
    #   osqp
    #   pyportfolioopt
    #   qdldl
    #   scikit-learn
    #   scs
    #   sentence-transformers
scs==3.2.7
    # via cvxpy
sentence-transformers==3.0.1
    # via -r requirements.in
sentencepiece==0.2.0
    # via gliner
setuptools-scm==8.1.0
    # via python-certifi-win32
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
spotipy==2.24.0
    # via -r requirements.in
sqlalchemy==2.0.36
    # via -r requirements.in
sqlite-vec==0.1.1
    # via -r requirements.in
sympy==1.13.2
    # via
    #   onnxruntime
    #   torch
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.7.0
    # via
    #   -r requirements.in
    #   openai-messages-token-helper
tokenizers==0.19.1
    # via transformers
torch==2.4.0
    # via
    #   gliner
    #   sentence-transformers
tqdm==4.66.5
    # via
    #   datasets
    #   gliner
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.44.2
    # via
    #   gliner
    #   sentence-transformers
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.2.2
    # via
    #   requests
    #   spotipy
    #   yt-dlp
websockets==13.0.1
    # via yt-dlp
wrapt==1.16.0
    # via
    #   pip-system-certs
    #   python-certifi-win32
xxhash==3.5.0
    # via datasets
yarl==1.17.1
    # via aiohttp
yt-dlp==2024.8.6
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
